mysite/requirements.txt
appdirs==1.4.4
asgiref==3.2.10
captcha==0.3
certifi==2020.6.20
chardet==3.0.4
chartjs==1.2
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1
django-ckeditor==6.0.0
django-cors-headers==3.7.0
django-cors-middleware==1.5.0
django-crispy-forms==1.11.0
django-debug-toolbar==3.1.1
django-debug-toolbar-force==0.1.8
django-dirtyfields==1.4.1
django-extensions==3.1.0
django-js-asset==1.2.2
django-makemessages-xgettext==0.1.1
django-modeltranslation==0.16.2
django-multiselectfield==0.1.12
django-nine==0.2.3
django-paypal==1.0.0
django-recaptcha==2.0.6
django-recaptcha3==0.4.0
django-rest-framework==0.1.0
django-specifications==0.4.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
filelock==3.0.12
flake8==3.8.4
future==0.18.2
game==0.0.5
gunicorn==20.0.4
idna==2.10
mccabe==0.6.1
numpy==1.19.5
Pillow==8.0.1
progressbar2==3.53.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
py==1.10.0
pycodestyle==2.6.0
pyflakes==2.2.0
pygam==0.8.0
pygame==2.0.1
python-gettext==4.0
python-utils==2.4.0
pytils==0.3
pytz==2020.1
requests==2.25.0
scipy==1.6.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.26.2
virtualenv==20.3.1
whitenoise==5.2.0
